docs: update README for modularize G2P
