+++ src/zh.ts
@@ -1,10 +1,8 @@
-/**
- * Preset for Chinese
- */
-
-import ChineseG2P from "./zh-g2p";
 import { useG2P } from "./g2p";
+import EnglishG2P from "./en-g2p";
+import ChineseG2P from "./zh-g2p";
 
+useG2P(new EnglishG2P());
 useG2P(new ChineseG2P());
 
-export * from "./index";
+export * from "./core";
